+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Latest changes
-## v2.7.6
-### Changes
-- Removed tooltips from tags on torrent details pages as they interfered with the native tooltip
+## v2.7.7
+### Features
+- Added a button that lets you filter all unfiltered results on the page with a single click. This requires individual upload hiding to be enabled in order to function. The feature is enabled by default, but you can easily disable it in Options > Settings > Filtering > Enable "Filter all unfiltered results" button.
+
+### Miscellaneous
+- Adjusted the padding of some buttons to reduce their five-heads
+- Adjusted the padding of the settings checkboxes
  
 ---